refactor: AdminController 클래스명 AdminApi로 변경 및 사용하지 않는 컨트롤러 메서드 삭제
